web/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
motor==3.7.1
pymongo==4.16.0
python-jose[cryptography]==3.3.0
httpx==0.28.1
pydantic-settings==2.6.1
python-dotenv==1.0.1
python-multipart==0.0.17
Pillow==11.0.0
google-generativeai==0.8.3
